Refactoring tess-calendar. Adding script to get next meeting link
